+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This version adds a dependency on Swift.
 
 - Properly demangle Swift class name (#2162)
 
+### Fixes
+
+- Errors shortly after SentrySDK.init now affect the session (#2430)
+
 ### Breaking Changes
 
 - Rename `- [SentrySDK startWithOptionsObject:]` to `- [SentrySDK startWithOptions:]` (#2404)